+++ test/local_album_test.dart
@@ -10,8 +10,10 @@ import 'package:local_image_provider/local_image_provider.dart';
 void main() {
   const String albumId1 = "album1";
   const String albumId2 = "album2";
+  const String albumId3 = "album3";
   const String title1 = "title1";
   const String title2 = "title2";
+  const String title3 = "title3";
   const int imageCount1 = 10;
   const int imageCount2 = 20;
   const String imgId1 = "img1";
@@ -20,12 +22,14 @@ void main() {
   const int height1 = 100;
   const int width1 = 200;
   const LocalImage coverImg1 = LocalImage(imgId1, creation1, height1, width1);
-  const LocalAlbum localAlbum1 =
-      LocalAlbum(albumId1, coverImg1, title1, imageCount1);
-  const LocalAlbum localAlbum1a =
-      LocalAlbum(albumId1, coverImg1, title1, imageCount1);
-  const LocalAlbum localAlbum2 =
-      LocalAlbum(albumId2, coverImg1, title2, imageCount2);
+  final LocalAlbum localAlbum1 = LocalAlbum(
+      albumId1, coverImg1, title1, imageCount1, LocalAlbumType.album.value);
+  final LocalAlbum localAlbum1a = LocalAlbum(
+      albumId1, coverImg1, title1, imageCount1, LocalAlbumType.album.value);
+  final LocalAlbum localAlbum2 = LocalAlbum(
+      albumId2, coverImg1, title2, imageCount2, LocalAlbumType.user.value);
+  final LocalAlbum localAlbum3 = LocalAlbum(
+      albumId3, coverImg1, title3, imageCount2, LocalAlbumType.shared.value);
   String requestedImgId;
   int requestedHeight;
   int requestedWidth;
@@ -33,7 +37,7 @@ void main() {
   LocalImageProvider localImageProvider;
 
   WidgetsFlutterBinding.ensureInitialized();
-  
+
   setUp(() async {
     List<int> imgInt = "087imgbytes234".codeUnits;
     imageBytes = Uint8List.fromList(imgInt);
@@ -61,11 +65,10 @@ void main() {
       expect(localAlbum1.coverImg, coverImg1);
       expect(localAlbum1.imageCount, imageCount1);
     });
-    test('Correct type returned', () {
-      expect(LocalAlbumType.fromInt(0), LocalAlbumType.all);
-      expect(LocalAlbumType.fromInt(1), LocalAlbumType.user);
-      expect(LocalAlbumType.fromInt(2), LocalAlbumType.generated);
-      expect(LocalAlbumType.fromInt(3), null);
+    test('albumType generated for album', () {
+      expect(localAlbum1.albumType, LocalAlbumType.album);
+      expect(localAlbum2.albumType, LocalAlbumType.user);
+      expect(localAlbum3.albumType, LocalAlbumType.shared);
     });
     test('hash works', () {
       expect(localAlbum1.hashCode, localAlbum1.hashCode);
@@ -89,6 +92,8 @@ void main() {
       expect(localAlbum1.title, roundtripAlbum.title);
       expect(localAlbum1.coverImg, roundtripAlbum.coverImg);
       expect(localAlbum1.imageCount, roundtripAlbum.imageCount);
+      expect(localAlbum1.transferType, roundtripAlbum.transferType);
+      expect(localAlbum1.albumType, roundtripAlbum.albumType);
     });
   });
   group('imageBytes', () {
@@ -99,6 +104,23 @@ void main() {
       expect(requestedHeight, height1);
       expect(requestedWidth, width1);
       expect(bytes, imageBytes);
+    });
+  });
+  group('LocalAlbumType', () {
+    test('correct type returned from int', () {
+      expect(LocalAlbumType.fromInt(0), LocalAlbumType.all);
+      expect(LocalAlbumType.fromInt(1), LocalAlbumType.album);
+      expect(LocalAlbumType.fromInt(2), LocalAlbumType.user);
+      expect(LocalAlbumType.fromInt(3), LocalAlbumType.generated);
+      expect(LocalAlbumType.fromInt(4), LocalAlbumType.faces);
+      expect(LocalAlbumType.fromInt(5), LocalAlbumType.event);
+      expect(LocalAlbumType.fromInt(6), LocalAlbumType.imported);
+      expect(LocalAlbumType.fromInt(7), LocalAlbumType.shared);
+    });
+    test('unknown ints convert to generic album type', () {
+      expect(LocalAlbumType.fromInt(8), LocalAlbumType.album);
+      expect(LocalAlbumType.fromInt(-1), LocalAlbumType.album);
+      expect(LocalAlbumType.fromInt(712), LocalAlbumType.album);
     });
   });
 }